+++ packages/backend/src/sbom/sbom-utils.ts
@@ -2,19 +2,24 @@ import { SbomPackage, Severity } from "shared-types";
 import { logger } from "../utils/logger";
 import licenseConfig from "./licenseConfig.json";
 import { isNpmPackage, isPythonPackage } from "./sbom-fetching-utils";
-import { Component, Dependency, LicenseInfo, NpmLicense } from "./sbom-types";
+import {
+  Component,
+  Dependency,
+  LicenseInfo,
+  NpmLicense,
+  RawPackage,
+} from "./sbom-types";
 
 function sortByLicenseLength(arr: LicenseInfo[]): LicenseInfo[] {
   return arr.sort((a, b) => b.license.length - a.license.length);
 }
 
-export async function getPackages(
+export function createSBOMPackages(
   dependencies: Dependency[],
   components: Component[],
   applications: Record<string, string>,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  pkgsInfo: any[]
-) {
+  rawPackages: RawPackage[]
+): SbomPackage[] | undefined {
   const sbomPackages: SbomPackage[] = [];
   const sortedLicenseConfig = sortByLicenseLength(licenseConfig);
   let filePath = "";
@@ -40,8 +45,9 @@ export async function getPackages(
           if (isPythonPackage(purl)) {
             registry = "PyPi";
             try {
-              const packageInfo = pkgsInfo.find(
-                (pkg) => pkg &&
+              const packageInfo = rawPackages.find(
+                (pkg) =>
+                  pkg &&
                   pkg.name === component.name &&
                   pkg.version === component.version
               );
@@ -56,8 +62,9 @@ export async function getPackages(
             }
           } else if (isNpmPackage(purl)) {
             try {
-              const packageInfo = pkgsInfo.find(
-                (pkg) => pkg &&
+              const packageInfo = rawPackages.find(
+                (pkg) =>
+                  pkg &&
                   pkg.name === component.name &&
                   pkg.version === component.version
               );
@@ -113,7 +120,13 @@ export async function getPackages(
 
   // Remove duplicates
   const sbomPackagesUnique = sbomPackages.reduce((unique, o) => {
-    if (!unique.some(obj => obj.packageName === o.packageName && obj.packageVersion === o.packageVersion)) {
+    if (
+      !unique.some(
+        (obj) =>
+          obj.packageName === o.packageName &&
+          obj.packageVersion === o.packageVersion
+      )
+    ) {
       unique.push(o);
     }
     return unique;
@@ -121,22 +134,21 @@ export async function getPackages(
 
   // Order SBOM packages
   const severityOrder = ["critical", "high", "medium", "low"];
-  sbomPackagesUnique.sort(
-    (a, b) => {
-      // First, order by severity
-      const orderRes = severityOrder.indexOf(b.severity) - severityOrder.indexOf(a.severity);
-      if (orderRes !== 0) {
-        return orderRes;
-      }
-      // Then order by package name
-      const packageNameRes = a.packageName.localeCompare(b.packageName);
-      if ([1, -1].includes(packageNameRes)) {
-        return packageNameRes;
-      }
-      // Otherwise use version number
-      return a.packageVersion.localeCompare(b.packageVersion);
+  sbomPackagesUnique.sort((a, b) => {
+    // First, order by severity
+    const orderRes =
+      severityOrder.indexOf(b.severity) - severityOrder.indexOf(a.severity);
+    if (orderRes !== 0) {
+      return orderRes;
     }
-  );
+    // Then order by package name
+    const packageNameRes = a.packageName.localeCompare(b.packageName);
+    if ([1, -1].includes(packageNameRes)) {
+      return packageNameRes;
+    }
+    // Otherwise use version number
+    return a.packageVersion.localeCompare(b.packageVersion);
+  });
   return sbomPackagesUnique;
 }